Updates to model objects
