+++ PersonApi/V1/Domain/Person.cs
@@ -21,6 +21,5 @@ namespace PersonApi.V1.Domain
         public IEnumerable<Language> Languages { get; set; }
         public IEnumerable<CommunicationRequirement> CommunicationRequirements { get; set; }
         public IEnumerable<PersonType> PersonTypes { get; set; }
-        public IEnumerable<ApiLink> Links { get; set; }
     }
 }
